Outline Kinect Proximity script
In order to debug in the SMILE lab, we need to be able to only take the nearest body on the kinect, as other people in the room interfere with the camera tracking just by being there. In order to do this, only the body which is closed to the sensor whould be recorded, or at least only that one should instantiate an object. Serval comments have been made in the BodySourceView script with suggestions of how this could be done.